+++ SourceCode/Services/Implementations/CountryService.cs
@@ -19,6 +19,6 @@ public sealed class CountryService
         if (principal is null) return Array.Empty<ListboxItem>();
         using var dbContext = Factory.CreateDbContext();
         var countries = await dbContext.Countries.ToListAsync();
-        return countries.AsEnumerable().Where(c => isAdministrator || principal.IsAuthorisedInCountry(c.Id)).Select(c => new ListboxItem(c.Id, c.EnglishName.Localized())).ToList();
+        return countries.AsEnumerable().Where(c => isAdministrator || principal.IsAuthorisedInCountry(c.Id)).Select(c => new ListboxItem(c.Id, c.EnglishName.AsLocalized())).ToList();
     }
 }